+++ src/components/OpinionsSection/opinion.module.scss
