+++ src/components/OpinionsSection/swiper.scss
@@ -1,0 +1,3 @@
+.swiper {
+  width: 1090px;
+}
