--- src/components/SearchFilter/search_filter.module.scss
@@ -1,74 +1,0 @@
-.form {
-  display: flex;
-  flex-direction: column;
-  padding: 1rem;
-}
-.label {
-  position: relative;
-}
-.error {
-  position: absolute;
-  left: 0;
-  right: 0;
-  top: -8px;
-  font-size: 12px;
-  color: red;
-  font-weight: bold;
-}
-.searchFilter {
-  &__top {
-    display: flex;
-    justify-content: space-between;
-    width: 100%;
-  }
-  &__bottom {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    width: 100%;
-    & > label {
-      margin: 1rem;
-      & > input,
-      & > select {
-        width: 280px;
-        height: 80px;
-        padding: 1rem;
-        background: linear-gradient(
-          108.1deg,
-          rgba(255, 255, 255, 0.4) 12.41%,
-          rgba(254, 254, 254, 0.1) 88.76%
-        );
-        box-shadow: 0px 4px 24px -1px rgba(0, 0, 0, 0.1);
-        backdrop-filter: blur(25px);
-        border-radius: 25px;
-      }
-    }
-  }
-  &__button {
-    border: none;
-    margin: 0;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    border-radius: 50%;
-    & > img {
-      width: 50px;
-      height: 50px;
-      margin: 15px;
-      object-fit: contain;
-    }
-  }
-  &__checkbox {
-    display: flex;
-    padding: 1rem 3rem;
-    &_active {
-      font-weight: bold;
-    }
-    & > span {
-      width: 100px;
-    }
-    & > input {
-      display: none;
-    }
-  }
-}
