+++ src/components/index.js
@@ -1,0 +1,10 @@
+export { Button } from "./Button/Button";
+export { Checkbox } from "./Checkbox/Checkbox";
+export { OpinionsSection } from "./OpinionsSection/OpinionsSection";
+export { DataList } from "./DataList/DataList";
+export { Footer } from "./Footer/Footer";
+export { Input } from "./Input/Input";
+export { Logo } from "./Logo/Logo";
+// export { Navbar } from "./Navbar/Navbar";
+export { SearchForm } from "./SearchForm/SearchForm";
+export { Select } from "./Select/Select";
