+++ src/modules/announcements/pages/Announcements.jsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from "react";
 import { useLocation } from "react-router-dom";
-import { SearchForm } from "../../../components/SearchForm/SearchForm";
+import { SearchForm } from "../../../components";
 import { ANNOUNCEMENTS_LIST } from "../../../constants/announcements";
 import { filterAnnouncements } from "../../../utils";
