+++ src/modules/auth/components/Form/RegisterForm/RegisterForm.jsx
@@ -2,8 +2,7 @@ import { useState } from "react";
 import { useForm } from "react-hook-form";
 import { yupResolver } from "@hookform/resolvers/yup";
 import * as yup from "yup";
-import { Button } from "../../../../../components/Button/Button";
-import { Input } from "../../../../../components/Input/Input";
+import { Button, Input } from "../../../../../components";
 import styles from "../form.module.scss";
 import { DUMMY_LOGINS } from "../../../../../mock/auth";
