+++ src/modules/main/pages/Main.jsx
@@ -1,7 +1,6 @@
 import { Section } from "../components/EncourageSection/Section";
 import dog from "../../../assets/image/encourages/dog.jpg";
-import { SearchForm } from "../../../components/SearchForm/SearchForm";
-import { OpinionsSection } from "../../../components/CustomerOpinions/OpinionsSection/OpinionsSection";
+import { SearchForm, OpinionsSection } from "../../../components";
 import { StepByStepList } from "../components/Steplist/StepList";
 
 export const Main = () => {
